requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-flask>=1.3.0,<2.0.0

# Code quality
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0

# Security
bandit>=1.7.0,<2.0.0
safety>=2.3.0,<3.0.0
